+++ src/main/java/com/_37coins/bcJsonRpc/pojo/Info.java
@@ -1,5 +1,7 @@
 package com._37coins.bcJsonRpc.pojo;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
@@ -10,16 +12,16 @@ public class Info {
 	private long version;
 	private long protocolversion;
 	private long walletversion;
-	private double balance;
+	private BigDecimal balance;
 	private long blocks;
 	private long timeoffset;
 	private long connections;
 	private String proxy;
-	private double difficulty;
+	private BigDecimal difficulty;
 	private boolean testnet;
 	private long keypoololdest;
 	private long keypoolsize;
-	private double paytxfee;
+	private BigDecimal paytxfee;
 	private String errors;
 	public long getVersion() {
 		return version;
@@ -39,10 +41,10 @@ public class Info {
 	public void setWalletversion(long walletversion) {
 		this.walletversion = walletversion;
 	}
-	public double getBalance() {
+	public BigDecimal getBalance() {
 		return balance;
 	}
-	public void setBalance(double balance) {
+	public void setBalance(BigDecimal balance) {
 		this.balance = balance;
 	}
 	public long getBlocks() {
@@ -69,10 +71,10 @@ public class Info {
 	public void setProxy(String proxy) {
 		this.proxy = proxy;
 	}
-	public double getDifficulty() {
+	public BigDecimal getDifficulty() {
 		return difficulty;
 	}
-	public void setDifficulty(double difficulty) {
+	public void setDifficulty(BigDecimal difficulty) {
 		this.difficulty = difficulty;
 	}
 	public boolean isTestnet() {
@@ -93,10 +95,10 @@ public class Info {
 	public void setKeypoolsize(long keypoolsize) {
 		this.keypoolsize = keypoolsize;
 	}
-	public double getPaytxfee() {
+	public BigDecimal getPaytxfee() {
 		return paytxfee;
 	}
-	public void setPaytxfee(double paytxfee) {
+	public void setPaytxfee(BigDecimal paytxfee) {
 		this.paytxfee = paytxfee;
 	}
 	public String getErrors() {
